backend/requirements.txt
Django==4.2
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1

